+++ libraries/transformer/src/main/java/androidx/media3/transformer/MuxerWrapper.java
@@ -19,6 +19,7 @@ package androidx.media3.transformer;
 import static androidx.media3.common.util.Assertions.checkArgument;
 import static androidx.media3.common.util.Assertions.checkNotNull;
 import static androidx.media3.common.util.Assertions.checkState;
+import static androidx.media3.common.util.Assertions.checkStateNotNull;
 import static androidx.media3.common.util.Util.areEqual;
 import static androidx.media3.common.util.Util.contains;
 import static androidx.media3.common.util.Util.usToMs;
@@ -47,7 +48,6 @@ import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.ScheduledFuture;
 import org.checkerframework.checker.nullness.qual.EnsuresNonNull;
 import org.checkerframework.checker.nullness.qual.MonotonicNonNull;
-import org.checkerframework.checker.nullness.qual.RequiresNonNull;
 
 /**
  * A wrapper around a media muxer.
@@ -99,6 +99,7 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
   private final String outputPath;
   private final Muxer.Factory muxerFactory;
   private final Listener listener;
+  private final boolean dropSamplesBeforeFirstVideoSample;
   private final SparseArray<TrackInfo> trackTypeToInfo;
   private final ScheduledExecutorService abortScheduledExecutorService;
 
@@ -113,6 +114,7 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
   private @MuxerMode int muxerMode;
   private boolean muxedPartialVideo;
   private boolean muxedPartialAudio;
+  private long firstVideoPresentationTimeUs;
 
   private volatile int additionalRotationDegrees;
   private volatile int trackCount;
@@ -125,16 +127,24 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
    * @param listener A {@link MuxerWrapper.Listener}.
    * @param muxerMode The {@link MuxerMode}. The initial mode must be {@link #MUXER_MODE_DEFAULT} or
    *     {@link #MUXER_MODE_MUX_PARTIAL}.
+   * @param dropSamplesBeforeFirstVideoSample Whether to drop any non-video samples with
+   *     presentation timestamps before the first video sample.
    */
   public MuxerWrapper(
-      String outputPath, Muxer.Factory muxerFactory, Listener listener, @MuxerMode int muxerMode) {
+      String outputPath,
+      Muxer.Factory muxerFactory,
+      Listener listener,
+      @MuxerMode int muxerMode,
+      boolean dropSamplesBeforeFirstVideoSample) {
     this.outputPath = outputPath;
     this.muxerFactory = muxerFactory;
     this.listener = listener;
     checkArgument(muxerMode == MUXER_MODE_DEFAULT || muxerMode == MUXER_MODE_MUX_PARTIAL);
     this.muxerMode = muxerMode;
+    this.dropSamplesBeforeFirstVideoSample = dropSamplesBeforeFirstVideoSample;
     trackTypeToInfo = new SparseArray<>();
     previousTrackType = C.TRACK_TYPE_NONE;
+    firstVideoPresentationTimeUs = C.TIME_UNSET;
     abortScheduledExecutorService = Util.newSingleThreadScheduledExecutor(TIMER_THREAD_NAME);
   }
 
@@ -256,7 +266,6 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
         checkArgument(existingFormat.sampleRate == format.sampleRate);
         checkArgument(existingFormat.initializationDataEquals(format));
       }
-      checkNotNull(muxer);
       resetAbortTimer();
       return;
     }
@@ -308,7 +317,8 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
    * @param data The sample to write.
    * @param isKeyFrame Whether the sample is a key frame.
    * @param presentationTimeUs The presentation time of the sample in microseconds.
-   * @return Whether the sample was successfully written. {@code false} if samples of other
+   * @return Whether the sample was successfully written, or dropped if configured to drop the
+   *     sample via {@code dropSamplesBeforeFirstVideoSample}. {@code false} if samples of other
    *     {@linkplain C.TrackType track types} should be written first to ensure the files track
    *     interleaving is balanced, or if the muxer hasn't {@linkplain #addTrackFormat(Format)
    *     received a format} for every {@linkplain #setTrackCount(int) track}.
@@ -328,12 +338,22 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
           presentationTimeUs,
           /* extraFormat= */ "%s",
           /* extraArgs...= */ canWriteSample);
+      if (firstVideoPresentationTimeUs == C.TIME_UNSET) {
+        firstVideoPresentationTimeUs = presentationTimeUs;
+      }
     } else if (trackType == C.TRACK_TYPE_AUDIO) {
       DebugTraceUtil.logEvent(
           DebugTraceUtil.EVENT_MUXER_CAN_WRITE_SAMPLE_AUDIO,
           presentationTimeUs,
           /* extraFormat= */ "%s",
           /* extraArgs...= */ canWriteSample);
+      if (dropSamplesBeforeFirstVideoSample
+          && firstVideoPresentationTimeUs != C.TIME_UNSET
+          && presentationTimeUs < firstVideoPresentationTimeUs) {
+        // Drop the buffer.
+        resetAbortTimer();
+        return true;
+      }
     }
     if (!canWriteSample) {
       return false;
@@ -343,8 +363,8 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
     trackInfo.bytesWritten += data.remaining();
     trackInfo.timeUs = max(trackInfo.timeUs, presentationTimeUs);
 
-    checkNotNull(muxer);
     resetAbortTimer();
+    checkStateNotNull(muxer);
     muxer.writeSampleData(
         trackInfo.index, data, presentationTimeUs, isKeyFrame ? C.BUFFER_FLAG_KEY_FRAME : 0);
     if (trackType == C.TRACK_TYPE_VIDEO) {
@@ -442,6 +462,12 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
   }
 
   private boolean canWriteSample(@C.TrackType int trackType, long presentationTimeUs) {
+    if (dropSamplesBeforeFirstVideoSample
+        && trackType != C.TRACK_TYPE_VIDEO
+        && firstVideoPresentationTimeUs == C.TIME_UNSET) {
+      // Haven't received the first video sample yet, so can't write any audio.
+      return false;
+    }
     if (!isReady) {
       return false;
     }
@@ -462,8 +488,8 @@ import org.checkerframework.checker.nullness.qual.RequiresNonNull;
     return presentationTimeUs - minTrackTimeUs <= MAX_TRACK_WRITE_AHEAD_US;
   }
 
-  @RequiresNonNull("muxer")
   private void resetAbortTimer() {
+    checkStateNotNull(muxer);
     long maxDelayBetweenSamplesMs = muxer.getMaxDelayBetweenSamplesMs();
     if (maxDelayBetweenSamplesMs == C.TIME_UNSET) {
       return;