+++ libraries/transformer/src/main/java/androidx/media3/transformer/Transformer.java
@@ -108,6 +108,7 @@ public final class Transformer {
     private boolean removeVideo;
     private boolean flattenForSlowMotion;
     private boolean trimOptimizationEnabled;
+    private boolean fileStartsOnVideoFrameEnabled;
     private ListenerSet<Transformer.Listener> listeners;
     private AssetLoader.@MonotonicNonNull Factory assetLoaderFactory;
     private AudioMixer.Factory audioMixerFactory;
@@ -148,6 +149,7 @@ public final class Transformer {
       this.removeAudio = transformer.removeAudio;
       this.removeVideo = transformer.removeVideo;
       this.trimOptimizationEnabled = transformer.trimOptimizationEnabled;
+      this.fileStartsOnVideoFrameEnabled = transformer.fileStartsOnVideoFrameEnabled;
       this.listeners = transformer.listeners;
       this.assetLoaderFactory = transformer.assetLoaderFactory;
       this.audioMixerFactory = transformer.audioMixerFactory;
@@ -320,6 +322,25 @@ public final class Transformer {
     @CanIgnoreReturnValue
     public Builder experimentalSetTrimOptimizationEnabled(boolean enabled) {
       trimOptimizationEnabled = enabled;
+      return this;
+    }
+
+    /**
+     * Set whether to ensure that the output file starts on a video frame.
+     *
+     * <p>Any audio samples that are earlier than the first video frame will be dropped. This can
+     * make the output of trimming operations more compatible with player implementations that don't
+     * show the first video frame until its presentation timestamp.
+     *
+     * <p>Ignored when {@linkplain #experimentalSetTrimOptimizationEnabled trim optimization} is
+     * set.
+     *
+     * @param enabled Whether to ensure that the file starts on a video frame.
+     * @return This builder.
+     */
+    @CanIgnoreReturnValue
+    public Builder setEnsureFileStartsOnVideoFrameEnabled(boolean enabled) {
+      fileStartsOnVideoFrameEnabled = enabled;
       return this;
     }
 
@@ -539,6 +560,7 @@ public final class Transformer {
           removeVideo,
           flattenForSlowMotion,
           trimOptimizationEnabled,
+          fileStartsOnVideoFrameEnabled,
           listeners,
           assetLoaderFactory,
           audioMixerFactory,
@@ -734,6 +756,7 @@ public final class Transformer {
   private final boolean removeVideo;
   private final boolean flattenForSlowMotion;
   private final boolean trimOptimizationEnabled;
+  private final boolean fileStartsOnVideoFrameEnabled;
   private final ListenerSet<Transformer.Listener> listeners;
   @Nullable private final AssetLoader.Factory assetLoaderFactory;
   private final AudioMixer.Factory audioMixerFactory;
@@ -769,6 +792,7 @@ public final class Transformer {
       boolean removeVideo,
       boolean flattenForSlowMotion,
       boolean trimOptimizationEnabled,
+      boolean fileStartsOnVideoFrameEnabled,
       ListenerSet<Listener> listeners,
       @Nullable AssetLoader.Factory assetLoaderFactory,
       AudioMixer.Factory audioMixerFactory,
@@ -787,6 +811,7 @@ public final class Transformer {
     this.removeVideo = removeVideo;
     this.flattenForSlowMotion = flattenForSlowMotion;
     this.trimOptimizationEnabled = trimOptimizationEnabled;
+    this.fileStartsOnVideoFrameEnabled = fileStartsOnVideoFrameEnabled;
     this.listeners = listeners;
     this.assetLoaderFactory = assetLoaderFactory;
     this.audioMixerFactory = audioMixerFactory;
@@ -926,7 +951,12 @@ public final class Transformer {
     if (!trimOptimizationEnabled || isMultiAsset()) {
       startInternal(
           composition,
-          new MuxerWrapper(path, muxerFactory, componentListener, MuxerWrapper.MUXER_MODE_DEFAULT),
+          new MuxerWrapper(
+              path,
+              muxerFactory,
+              componentListener,
+              MuxerWrapper.MUXER_MODE_DEFAULT,
+              /* dropSamplesBeforeFirstVideoSample= */ fileStartsOnVideoFrameEnabled),
           componentListener,
           /* initialTimestampOffsetUs= */ 0);
     } else {
@@ -1114,7 +1144,8 @@ public final class Transformer {
             checkNotNull(outputFilePath),
             muxerFactory,
             componentListener,
-            MuxerWrapper.MUXER_MODE_DEFAULT),
+            MuxerWrapper.MUXER_MODE_DEFAULT,
+            /* dropSamplesBeforeFirstVideoSample= */ false),
         componentListener,
         /* initialTimestampOffsetUs= */ 0);
   }
@@ -1144,7 +1175,8 @@ public final class Transformer {
                     checkNotNull(outputFilePath),
                     muxerFactory,
                     componentListener,
-                    MuxerWrapper.MUXER_MODE_MUX_PARTIAL);
+                    MuxerWrapper.MUXER_MODE_MUX_PARTIAL,
+                    /* dropSamplesBeforeFirstVideoSample= */ false);
 
             startInternal(
                 TransmuxTranscodeHelper.createVideoOnlyComposition(
@@ -1193,7 +1225,8 @@ public final class Transformer {
             checkNotNull(oldFilePath),
             muxerFactory,
             componentListener,
-            MuxerWrapper.MUXER_MODE_DEFAULT),
+            MuxerWrapper.MUXER_MODE_DEFAULT,
+            /* dropSamplesBeforeFirstVideoSample= */ false),
         componentListener,
         /* initialTimestampOffsetUs= */ 0);
   }
@@ -1273,7 +1306,8 @@ public final class Transformer {
                     checkNotNull(outputFilePath),
                     muxerFactory,
                     componentListener,
-                    MuxerWrapper.MUXER_MODE_MUX_PARTIAL);
+                    MuxerWrapper.MUXER_MODE_MUX_PARTIAL,
+                    /* dropSamplesBeforeFirstVideoSample= */ false);
             if (shouldTranscodeVideo(
                     checkNotNull(mp4MetadataInfo.videoFormat),
                     composition,